Fixed bug

+++ Binomial.java
@@ -7,9 +7,9 @@ public class Binomial {
 		// System.out.println(binomial1(Integer.parseInt(args[0]),
 		// Integer.parseInt(args[1])));
 
-		// int n = 52;
-		// int k = 5;
-		// System.out.println(binomial1(n, k));
+		// int n = 50;
+		// int k = 25;
+		// // System.out.println(binomial1(n, k));
 		// System.out.println(binomial(n, k));
 
 		// Testing the optimized binomial implementation:
@@ -29,11 +29,11 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 		//// This function creates a 2D array, say memo, and then initializes all its
 		/// elements to -1. It then calls binomial(n, k, memo), which does all the
 		/// heavy lifiting.
-		int[][] memo = new int[n + 1][k + 1];
+		long[][] memo = new long[n + 1][k + 1];
 		for (int i = 0; i < n + 1; i++) {
 			for (int j = 0; j < k + 1; j++) {
 				memo[i][j] = -1;
@@ -42,7 +42,7 @@ public class Binomial {
 		return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
